web/requirements.txt
7quart==0.19.4
aiosqlite==0.19.0
python-dotenv==1.0.0
discord.py==2.3.2
